refactor(e2e): split login tests into separate files
- Split monolithic 01-02-login.test.ts into 4 focused test files:
  - 01-02-01-login-valid-credentials.test.ts
  - 01-02-02-login-invalid-credentials.test.ts
  - 01-02-03-login-empty-fields.test.ts
  - 01-02-04-login-navigate-to-registration.test.ts
- Each file tests a single scenario for better isolation
- Added proper expect assertions to fix playwright/expect-expect lint
